requirements.txt
fastapi==0.95.0
uvicorn==0.23.1
pydantic==2.1.1
motor==3.1.0
python-dotenv==1.0.0
